requirements.txt
asttokens==3.0.0
backcall==0.2.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
hpack==4.0.0
huggingface-hub==0.26.5
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.7.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prettytable==3.11.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-box==7.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
tzdata==2024.2
wcwidth==0.2.13
zipp==3.20.2
zstandard==0.23.0